Related with issue 29064: Added test for the new NoteDataSource
A test for making sure that the new manual data source for the notes works properly.
It does the fetch, add and remove methods of the NoteDataSource.java.
The testFetchNotes test has a user of an American organization which creates a new note, so the note's
organization is the same American organization. After that the organization of the note is changed to
another one which is a Spanish organization for making sure that the two organization are divergent.
After that a fetch is done, now, the user has no access to the new organization of the note, but it have
to be possible fetching the note. This is what this test does, a fetch making sure that the filter by the
organizations is not done.
The class DatasourceTestUtil has been modified for the correct performance of the testRemoveNotes.