fixes issue 29594: ShowAcct preference not working on 2nd level accounting tabs
The ShowAcct preference value was not being retrieved properly in child tabs in the addChildView method, because the windowId was undefined.
This is because windowId is only set for the top view in the hierarchy. Now the problem is solved because the windowId value is retrieved using the reference to the standard window.

--HG--
extra : rebase_source : ab9a1d8b78db39dc86b389783d312711f444f96b