Fixes bug 29666: Fix NPE in Costing Background with Standard Costing Rule
In Cost Adjustment project new method has been added to calculate an opening inventory cost. This method uses 'CostinUtils.getStandardCostDefinition' method, that tries to get an standard cost for that opening inventory. If it does not found any standard or legacy cost, it returns null. So, in this moment a NPE error was happening. To fix this problem, at this point a controlled error is raised to notify the user that no cost has been found.