Fixed issue 29705: Description field is set with "null" in the FA Transaction
Description field is filled in with "null" in the FA Transaction when matched
to GL Item.

The problem was in the Match Statement process, more specifically, in the
Add new transaction button. The AddTransanctionActionHandler didn't
take in consideration that the received description could be null.

The fix for this issue is to check if the received description is null
or blank.

Now, the description field fill in the correct description.