fixes issue 29665: NP Error raised when using HQL Based Tables on a level 2 tab
A unnecessary request was being done when accessing to the HQL query based subtab. This was caused because the tab was not being identified properly as a subtab.
The view.parentProperty was used to identify if we are on a child tab. This property could be undefined depending on the configuration of the tab in the application dictionary.
Now view.parentView is used instead of view.parentProperty because it does not return undefined in this cases

--HG--
extra : rebase_source : 7be0fa11315de6c94a5608816b9a3f62a545515e