fixes issue 29667: NPE Error when a child tab is based on an HQL Query table
The NPE was raised because the code was retrieving the entity for a table not based on a db table. This kind of tables do not have BaseOBObjects.
The FIC was already controlling this when the tab has a related table of this type but it was not considering the case when the parent table is related to a table of this type.