Fixed bug 29035:Problems with inherited permissions in process definition
The problem was that when a process containing parameters defined as "window" is
launched , this manual role has not access to windows contained in that process
definition.

The cause of this issue is that before 14Q3, no security check was done on P&E
grids, so data always was retrieved.From 15Q1, security is checked requiring
explicit access to P&E grid.

The issue is fixed by inheriting access from the process, this is if the process
is accessible the grid within the P&E doesn't require to have explicit access
but inherits from the process itself.