Related to issue 14631: merge of Openbravo.properties in a legible way
The problem was when an update of core include a new property in
Openbravo.properties.template.  "mergeOpenbravoProperties()" method
merges Openbravo.properties with Openbravo.properties.template.

As a result of this merge, Openbravo.properties was working properly but in
a not very legible way.

It has been fixed this problem adding a new method that added new properties
in a legible way.