Fixed issue 28908: Not possible to delete a variant if has an image
Not possible to delete a variant if the generic product has an image.

The problem was that the image is shared among the generic product and
its variants.

The solution is to check if other products has the same image_id when deleting
 a record. In case that this check is true, the image isn't deleted from the
database, otherwise, the image is deleted from the database.