Fixes issue 29319: Performance problems validating a Costing Rule
Query in checkTransactionsWithMovDateInClosedPeriod method of CostingRuleProcessOnProcessHandler class, which has performance problems, has been changed.
Now M_Transaction.movementdate are truncated and grouped in order to have only one record per day instead of one record per transaction, which enhance a lot the performance of the query.
The query has been created in XSQL, because subquerys in FROM statement in HQL are not supported.