Related to issue 29496: Not possible to do a return with non stocked product
Not possible to do return to vendor using non stocked product.

The problem here was that the Return to Vendor line of a non stocked product
appears in the Pick/Edit Lines window in the Return to Vendor Shipment
window, but, when selected, a error message appears and it's not possible
to done the picking.

Now, we save the property stockable of the product, and we check if the product
is non stockable. Now, the Pick/Edit Lines works fine and it is possible to
process the Return to Vendor Shipment.