Related with bug 29612:New grid configuration to improve population of fk combo
The new grid configuration allows to change the way the foreign key filter drop downs are populated. Usually the rows shown in a fk filter drop down are only those that are being used in the referencing table. But that can be very slow if the referenced table has only a few records and the referenced tablas has a huge volume.

To address this problem, this new grid configuration allows to show all the records of the referenced table, even if they are not being used in the referencing table. When the queryService is being created to populate the filter drop down, the main entity is changed to the referenced table, and the criteria is updated. Also the where clause of the referencing table is now ignored.