Fixed bug 29823: Open/Close Period Control shows calendars associated to organizations
The Open/Close Period Control window was showing all the periods available at the C_Period table.
The target of this window is to open/close periods in fiscal calendars, so it should only show periods belonging to a fiscal calendar linked to an organization set as ready.

The fix has 2 parts:
1. Added a hql where clause to include only periods belonging to a fiscal calendar linked to an organization set as ready.
2. The hql filter clause was wrong because it was filtering by the c_period.ad_org_id. Instead it should be filtering by the organization linked to the calendar's periods