Related with bug 29612: Applies properly the client/organizations filter
If the column being filtered referenced the Client or Organization entity, then the resulting query was not including the where clause to filter the readable clients and organizations, respectively. This is a known issue, reported here [1].

This issue cannot be easily fixed, so this project address it by adding manually the readable client or organization filter if the entity being filtered is client or organization and if the query is being built as a result of opening the dropdown of one of the new unfiltered foreign key filters.

[1] https://issues.openbravo.com/view.php?id=29846