Fixed issue 29847: Applied bad format number when running Payment Report in PDF
Decimal and Grouping separators defined in Format.xml are not applied when
running Payment Report in PDF.

The solution is to fix the Payment Report number format in the jrxml file.
Also, a new parameter must be added to the Report.

Now, all the number fields of the report are formatted as the Format.xml
definition.