Fixes issue 29896: Prevents unneded datasource request when filtering
When the grid was filtered in a scrolled down grid, two datasource requests to load the grid were bein
g done instead of one. This happened due to two reasons:
- The function scrollTo of the grid's body was being invoked while the grid was being filtered. This f
unction can be skipped in this case because when a datasource request is triggered due to a filter eve
nt, the grid is always automatically scrolled up to the top
- If the OBViewGrid.getRecord function was invoked in the middle of a datasource request, an additiona
l request was done. To prevent this, in this case the function returns Array.LOADING

To detect if the grid is in the process of doing a datasource request, the isFilteringExternally flag
is used. This flag is set in the handleFilterEditorSubmit function, and only if the change in the filt
er editor is going to trigger a datasource request. The flag is reset in the callback of that function
.

--HG--
extra : rebase_source : f56ae456262593cc04e09c9d8db5f52f2cafb53e