Fixes issue 29896: Prevents unneded datasource request when filtering
When the grid was filtered in a scrolled down grid, two datasource requests to load the grid were being done instead of one. This happened due to two reasons:
- The function scrollTo of the grid's body was being invoked while the grid was being filtered. This function can be skipped in this case because when a datasource request is triggered due to a filter event, the grid is always automatically scrolled up to the top
- If the OBViewGrid.getRecord function was invoked in the middle of a datasource request, an additional request was done. To prevent this, in this case the function returns Array.LOADING

To detect if the grid is in the process of doing a datasource request, the isFilteringExternally flag is used. This flag is set in the handleFilterEditorSubmit function, and only if the change in the filter editor is going to trigger a datasource request. The flag is reset in the callback of that function.

This previous fix did not work properly, as the criteria passed to the willFetchData function was not proper, it needed to the processed in the convertCriteria function.

--HG--
extra : rebase_source : 48cb2716519f1a7952f404f663c46c69bd5be993