Fixes bug 29986: Correct trx status assigned while deleting from reconciliation
while checking or unchecking lines in Reconcile popup 'updateTransactionStatus' method is called to assign the proper status to the transaction. While unchecking 'Withdrawn not Cleared' or 'Deposited not Cleared' statuses has to be assigned to the transactions. This part is was wrong. When the transaction was of BP Deposit 'Withdrawn not Cleared' was being assigned and when the transaction was BP Withdrawal 'Deposited not Cleared' was being assigned.

--HG--
extra : transplant_source : %FCt%C9%21%FC%8C%0D%F6%F9%91%D2i%AD%98-%DF%89%DA-%AF