Fixes issue 29981: NPE in Add Payment process request
When selecting a business partner with the magnifying glass in Add Payment process request, which does not match to the type of the document (a Vendor when the type is Received In); the payment method of the business partner was null and when accessing to its id, a NPE was raised.
Now, if payment method is null will be checked.