Fixed issue 29641: With lazy filtering, summary functions display wrong value
The problem is when a summary function field exists, in this case, it is not taken into account the
fact of having the lazy filtering enabled. So, when a summary function is added to a field using the UI,
the values of that summary function are displayed before clicking the check icon, under the funnel icon,
in the right.
For fix this, a condition that checks if the lazy filtering has been enabled has been added.