Fixed issue 25278: Long time to update a warehouse with thousends of locators
On an environment with 60.000 locators for one of the warehouse, when updating
a field of the warehouse it takes a long time to save the record.

Part of the problem is the update of the m_warehouse table. When updated, the
trigger launch an update of the organization id for M_LOCATOR and
M_WAREHOUSE_ACCT tables.

The solution is to check if the organization has changed and, if not, avoid the
unnecessary update of those tables.