Fixed issue 29638: Google Calendar widget is not working properly.
In the workspace, adding a google calendar windget was not working properly. It
appeared blank.

It is necessary to set the field "email" of the user, in the user window, for being able
to show the asociated calendar in the widget. It is also necessary to be logged in the same
google account in the same browser for being able to display the calendar.

The problem was that it was not correctly displayed at first and it was necessary to refresh
the workspace for making the Google Calendar work properly.
This was because the email of the user was not correctly set.

The "evaluateContents" function has been overwritten in then "ob-url-widget.js" for granting
its correct behavior.