Fixed issue 29709: Physical inventory cannot be created.
The problem was that if it was a very small negative Book Quantity number, it was not
possible to create a Physical Inventory line.

It was because the "ScientificToDecimal" function was not working properly with negative
numbers. The '-' sign was taken as a number instead of as a sign, that was the reason of
obtaining a value like 0.000-3655 when -0.0003655 was expected.

To fix this, a condition has been added which takes into account the fact of having
the '-' sign and to deal with the sign as a sign and not as a number.