Fixed issue 30014: There are problems with big numbers in scientific notation
The problem was in OB.Utilities.Number.OBMaskedToOBPlain.
When the function was executed with a big negative number in scientific notation or a positive number with the "+" sign at the beginning, the
obtained decimal value was not correct.
This was caused because the length of the variable "number" (which was the number in scientific notation), was used for creating the final number instead of the
length of the variable "plainNumber", (which was the number in decimal notation).
This generated a problem because, when the spected result was a number longer than the number in scientific notation, it was not correctly obtained.
For example:

-1.0564E7  returned  -10564000.  This is correct, but:

-1.0564E8 returned   -10564000.  And this is not correct.

--HG--
extra : rebase_source : 3f986b66a707d11d7b30d26a12e73980703bd3cc