Fixed issue 29857: When there are not linked items it appears "Loading"
When there are no linked items, the linked items category section shows the message "Loading data..."
instead of the message "No items to show".

The problem was that the fact of not having any item category was not correctly handled.
In the initWidget function, when the linkedItemCategoryListGrid is created, in the function "dataArrived"
there was a condition which called the function that showed the correct message "No items to show" when the
startRow was '0' and when de endRow was '-1', the problem was that in this case of not having any linked Item
category, the endRow and the startRow where both '0', so, the correct message was never shown and it always shows
the message "Loading data...".
To fix this, the condition has been changed, now it is taking into account the length of the data, and if it
is '0' that means that there are not linked items so, the message "No items to show" is shown.