Fixed issue 30032: Pay button is enabled in the Receipt discounts tab
The problem was that the Synchronization Helper was sending 'synchronized' events while the discounts tab was open, enabling the button.
A check has been added to disable the button (or keep it disabled) if a request to disable it has been done, or if the this.isEnabled attribute is false (meaning that the toolbar has been disabled).

--HG--
extra : rebase_source : 4ebbd67247728e9fb45d967eeab8ffdb61da2e56