Fixed issue 30047: Finance role should be able to select a account in G/L Report
In an organization structure similar to F&B, a role with access to "US"
organization should be able to filter by specific accounts in General Ledger
Report, although account tree is defined at F&B Organization.

The problem was that the F&B US Finance role cannot select none of the accounts
defined for F&B International group.

The solution is to grant access to the accounts defined for F&B International
group. It has no sense that the user cannot select those accounts, but the user
can see the report with all it's accounts.