Fixed issue 30037: Completed quantity in the Work Requirement is not updated
Completed quantity in the Work Requirement is not updated under some
circumstances.

The problem was that the ma_workeffort_validate wasn't updated the UPDATED
column of the ma_wrpahse table and you can save even if the line had changed in
the work effort window.

The solution is to update the column with the actual date. Now, when you try to
save it, a error appears in the grid and the save is canceled.

--HG--
extra : rebase_source : 8ddd7f5fba865d8802e8ebc1e9508bfe06a3bbb5