Fixed issue 29946: Inactive records cannot be retrieved using dal WebServices.
It is possible to get the inactive records using JSON WebServices for example, as:

http://localhost:8080/openbravo/org.openbravo.service.json.jsonrest/Country?_noActiveFilter=true

But it is not possible to get the inactive records using dal WebServices.
As it would be interesting to have the choice of showing the inactive records too, as it is possible in the
JSON WebServices, a condition which checks if the "_noActiveFilter" has been included in the "DalWebService"
class.
Now it is possible to chose if one wants to get the inactive fields or not, as in the JSON WebServices:

http://localhost:8080/openbravo/ws/dal/Country?_noActiveFilter=false  // The inactive records are not shown.

http://localhost:8080/openbravo/ws/dal/Country?_noActiveFilter=true   //The inactive records are shown.