Related with 29946: Some tests have been performed.
Some test cases have been performed for testing the correct behaviour of apllying the
"_noActiveFilter" to the dal and JSON WebServices.
There are four tests. Two per WebService, one of them is putting the "_noActiveFilter" to true,
it works if it gets the inactive records in the JSON WebServices and also in dal WebServices.
The other test is putting the "_noActiveFilter" to false, it works if it don't get the inactive
records.