Fixes issue 30063: Grid is properly loaded after refreshing with a selected row
The problem was caused by this code (the this attribute is a ResultSet), which is contained in the fetchRemoteData function:

      } else if (this.grid.refreshingWithSelectedRecord) {
        // if the grid was refreshed with a record selected, use the range that contained that record
        //  instead of using targetRecordId to improve the performance
        startRow = this.grid.selectedRecordInitInterval;
        endRow = this.grid.selectedRecordEndInterval;
      }

If the grid is refreshed while one of its records is selected, then the startRow and endRow will be set so that the requested page contains the selected record. The problem was that the ResultSet.localData attribute was not being properly set, as at this points it was expected to contain the 'loading' value for all the rows that are being requested. As a result of this the localData attribute became malformed when the response was processed, and this caused the grid to misbehave.

To fix this, the localData is configured to wait for the proper records:

        this.localData = [];
        this.setRangeLoading(startRow, endRow);

--HG--
extra : rebase_source : efd3af5f4d4114f4bab3a0998e3afed3f9c77958