Fixed issue 29863: Title of process buttons are not showing the correct text.
When a process button allows multiple selection, the text shown in the button was not correct.

The problem was that the code for handling the change of the text inside a process button, that allows multiple selection,
didn't exist before.
To fix this the code needed to deal with the change of the name of the button for multiple selection has been  created.