Related to issue 29953: specific flow was not working fine
When a quotation is reactivated without load it the flow was not working because the order id of the quotation was being set to null.
Now id is not set to null, so when the quotation is reactivated we have the id to be used to set the oldid in the new reactivated quotation.

Creation date is updated because documentNo also is updated.

One popup is shown if the user try to reactivate a quotation whitout ID