Fixed issue 30112: The message shown trying to save a datapackage isn't correct
The problem was that the message associated to the label "20102" was not correct in this case.
To fix this, a new message has been added to AD_MESSAGE, and now, the message
'The java package of the Data Package must be within the java package of its module.' is
shown when the java package of the module is not contained in the java package of the data
package.
Also, the duplicated where clause has been removed from the trigger.

This fix has been tested with postgres and also with oracle.