Fixed issue 30116: Outstanding amount of sales order payment plan is not updated
The problem was that the outstanding amount of sales order payment plan is not
updated when payment method is set as automatic receipt.

Before the commit which introduces this regression,  the function checked the
status of the payment. With the changes,  the function checks if the payment
schedule detail is paid.

The problem was in the update of the payment plan of the sales order,
specifically in WHERE clause for outstanding amount.

Now, the where clause works as expected, and the outstanding amount is updated
properly.

--HG--
extra : rebase_source : f76cc41351f214e5134b312091ede540b286a5d2