Fixed issue 30121: Payment plan is updating wrong in some circumstances
Payment plan is updating wrong if Payment Method is set as Automatic Receipt
and a sales order is partially paid.

The problem was that the function duplicates the value of the received amount
created in the Complete process of the invoice, because the function searches
all the payment details of the payment plan that has an invoice paid
(invoicepaid column).

The solution is to search those payment details that are not pre-paid and then
update the payment plan.

Now, the values are filled correctly.

--HG--
extra : transplant_source : w%19%97%81%9D%2Bzz%99%BC%A7%89%1A%C5%F3%99%80%BD%2A%88