Fixed issue 30092: the bo.xslt template is working wrong.
The problem was that it was not possible to get the correct url by clicking
in the links. For example, if you went to the url:

https://livebuilds.openbravo.com/erp_main_pgsql/ws/dal/ADUser?template=bo.xslt

And you click in some link, the obtained url was:

https://livebuilds.openbravo.com/erp_main_pgsql/ws/ADUser/100?template=bo.xslt

instead of:

https://livebuilds.openbravo.com/erp_main_pgsql/ws/dal/ADUser/100?template=bo.xslt

The problem was in the bo.xslt template, in the href variable. It was being set wrong in some cases,
it removed one extra level from the url. This is OK when the url has the id after the entity name, like:

https://livebuilds.openbravo.com/erp_main_pgsql/ws/dal/ADUser/100?template=bo.xslt

But in the case of not having the id after the entity name in the url, it is wrong to substract a level
from the url.

In this case, that does not have the id after the entity name, didn't work well:

https://livebuilds.openbravo.com/erp_main_pgsql/ws/dal/ADUser?template=bo.xslt

The substraction of the level retuned a bad url, the "/dal" is missing:

https://livebuilds.openbravo.com/erp_main_pgsql/ws/ADUser/100?template=bo.xslt

To fix this the bo.xslt has been changed, now it takes into account if the url contains the id after
the entity name. If it does, a level is substracted. If not, no.

The WebServiceUtil.applyTemplate method has been changed too. Now the parameter hasId is set here and is
passed to the bo.xslt. This boolean parameter is true if the url has the id after the entity name,
and false if not.