Fixes issue 30132: Product Transactions tab not showing records after filtering
The problem happened on the client filtering. In this case, the id of the property that references the parent tab is included in the criteria (using the convertCriteria function of the OBViewGrid class).
For those tabs which does not have any property (column) marked as link to parent, the parent property was not being requested to the datasource and when applying the criteria on the local filter the value for the parent property in every record was undefined. That was the reason why the filtering was not returning any record.
Now if the tab does not have any property set as link to parent, we look for the property pointing to the parent tab and include it in the selected properties returned by the datasource.

--HG--
extra : rebase_source : 126d25fcbeb9b04ed856c451b025e080c46d3f82