fixed issue 28911: Possible to create new rows in edit only tabs in grid view
The problem was that when the UI pattern of the tab was set as "edit only", it was
possible to create new records in grid view, after editing the last record in grid view
and pressing the tab key and also by pressing the down arrow key.

The problem was in the ob-standard-view.js. There were two functions, one called
"setEditOrDeleteOnly" and another one called "setSingleRecord".
The first one, was calling another function called "setListEndEditAction" that changed
the value of "listEndEditAction" from "next" to "done" and that is the correct behaviour,
because if it is next, it allows to create a new row and in these cases we don't want it.
So, in the second function, "setSingleRecord", it was necessary to call the same function
for changing value of "listEndEditAction" from "next" to "done" for avoiding the bad behavior
of being able to create new rows.

--HG--
extra : rebase_source : 64b6c3b1c936578cef1b364323150d645491c59c