Related to issue 30090: displayed value should be set in fk fields
When the default value of a foreign key field in a new record does not exist already in the grid data, the display value can not be recovered in the setEditValue() method.
In order to ensure the display value is updated to reflect the new edit-value for this field, the display value is now explicitly set.
Also, we send null ROW_ID to avoid the edit values be overriden by the FIC response