Fixes issue 30189: Values could be overwritten in Modify Payment Plan
In the Modify Payment Plan grid an add function is creating an object to set the values for the new records. This function is setting an empty string for the id property, i.e., every new created record is generated with the same id. This causes that the selection record change is not handled properly under some circumstances in the _startEditing method of the ListGrid class.
To avoid this problem, the id property is not included in the object instead of leaving it empty.