Fixed issue 29939: Only readable Sales Representatives are loaded in Web POS
Before the fix, web POS was filtering organizations based just on the user.
Having the case where the user is defined in a readable organization but the BP
is defined in a non-readable organization, WEB POS will load this business
partner as a sales representative.

This fix adds a filter to the query to retrieve just this sales representatives
which are defined in readable organizations, It means that BP and User entities,
should be defined in readable organizations.