Fixes issue 30223: delete popup is not hidden under some circumstances
The problem was that if two or more records were selected and deleted, the records were deleted properly but the "Deleting..." popup is not hidden. The user will not be able to interact with the window until he refreshes it. This happends when in a grid that contains a tree grid view

This happened when a grid contains a treegrid and this one is not open first. In that case this element "view.treeGrid.data.handleUpdate" has not been created yet.

To avoid the problem, it has added a condition that checks whether the properties have been started correctly and thus, if the properties can be accessed.