Fixes issue 29612: FK filters with drop down disable behave like text filters
If the "Disable Foreign Key Combo" grid configuration is enabled, then the filters of foreign key fields will not include a dropdown. In this case the filter will look like a standard text filter, so it ha
s been decided that is should act like one. This mean the following grid properties should work:
- Text Field Filter Behaviour. Allows to specify the default operator, no change was needed
- Threashold to Trigger Filter on Text Field. Prevents executing several requests to the datasource in short time span. No change was needed
- Filter On Change: Allows to specify if a filter datasource request should be done when the user types, before the filter editor loses the focus.

The changes done in this changeset fix the third case. The actOnKeypress property is set to true if the FilterOnChange grid config is enabled, otherwise the filter was only done when the filter editor lost
 the focus. For this to work properly, another change was needed: the symbol of the default operator should not be shown in the filter editor.