Fixes issue 30206: Wrong payment status under some circumstancies
comingFrom internal parameter in FIN_PaymentProcess.java was always setted to "TRANSACTION" instead of using received comingFrom value, when Automatic Deposit/Withdrawn was activated and execution type was Automatic and not Deferred.
Because of that, transaction was not created when creating a payment from Payment In or Payment Out with this configuration.

--HG--
extra : rebase_source : e441643d3a975e0e33ff96dd4d35362451c4f533