Fixes issue 30230: Wrong payment status under some circumstancies
comingFrom internal parameter in FIN_PaymentProcess.java was always setted to "TRANSACTION" instead of using received comingFrom value, when Automatic Deposit/Withdrawn was activated and execution type was Automatic and not Deferred.
Because of that, transaction was not created when creating a payment from Payment In or Payment Out with this configuration.

--HG--
extra : transplant_source : r%DA%F3L%C8%BE%B6u%B7%F4%E9v%1C%D5%A9%81NP%87p