Fixed issue 29978: Enable tree reference selectors in Pick & Exectute windows.
It was not possible to use tree reference selectors in the pick & execute windows.

To enable this the "ob-formitem-tree-filter.js" has been changed.
Before, there was a function which copied the functions from the view grid, the problem was that this
doesn't exist for the Pick and Execute windows.

Now, it takes into account the pick and execute windows, this has been done by a
condition which checks if "paramWindow" is true, which means that is a pick and execute window.

A new function has been developed too, for granting the correct behavior of this new functionality.
This function, copies some functions from the grid of the pick and execute windows.