Fixed issue 29463: It is possible to pay a return ticket if there is no cash
It is possible to give change or pay a return ticket if there is no cash
available.

The solution is to check if there is cash to cover the overpayment. If no, the
cash payment is deleted an a warning message appears indicating that there is
not enough cash and that the cash payment has been deleted.

--HG--
extra : rebase_source : 182246a87d15d7a6a33c30fa701be1c83926278d