Fixed bug 30143: Add Multiple Payments in Financial Account
Due to the Add Transaction refactor, the ability to create transactions from several payments in a batch was missed.
This development adds a new button in Financial Account window (Add Multiple Payments) where the user must enter a transaction date and select one or more payments. The system will create a transaction per selected payment and it will process all of them. If any exception is raised at any point in time, no transactions will be created.

The payments shown in the P&E are exactly the same as in the payment selector at Transaction tab.
The great advantage of this process is that you can select multiple payments at the same time, and that the transactions are automatically processed, so it is very agile from the user point of view.

--HG--
extra : rebase_source : e5b93976478f776727cf20d1bb90671975ac258f