Fixed issue 30133: The date shown in the cash up ticket could be different
The problem was that the date shown at the end of the cash up process could
be different than the date saved in the backend.

The solution is to update the time field of the ticket before it's printed, and
update the date shown in the final step of the cashup.

Now, the date shown in the cash up ticket and the cash up backend are the same.

--HG--
extra : rebase_source : fd89903b8a01ce6354a47846264e1fee9628825e