Fixes 30135: On windows the ant setup tool writes incorrect source.path
The ant setup tool was writing an incorrect source.path because the user.dir property obtained under Windows contains backslashes as file separator in the path.
Now this case is controlled, and in case we are under a Windows operative system, the backslashes are replaced with a slash. The slash can be safely used in all platforms because Java internally can handle it and do the translation into platform-specific paths

--HG--
extra : rebase_source : 109a418d734ab59c062d7aaa0533beffc93f96c6